Harden Lot 5 security contract boundary
Add the runtime-free andromeda-security-contract crate with canonical surface, permission family, permission id, plane-binding, and policy evidence vocabulary.

Wire QUIC Procedure gateway validation to the canonical execute permission id, reject non-canonical execute aliases before dispatch, and replace surface-scope ordinal comparisons with typed enum equality in QUIC connection and gateway state.

Add topology and doctrine guards keeping the security contract crate below core/contract/proto/observe/quic/runtime/durable crates, document DEC-041, and cover duplicate manifest permission id/family drift.

+++ crates/andromeda-quic/src/connection/state_machine.rs
@@ -90,7 +90,7 @@ impl Connection {
 
         // Validate that the certificate's surface scope matches the connection plane.
         let required_scope = crate::mtls_identity::plane_to_required_surface_scope(self.plane);
-        if identity.surface as u8 != required_scope as u8 {
+        if identity.surface != required_scope {
             return Err(protocol_error(
                 "certificate surface scope does not match connection plane",
             ));
